New rx_agc implementation (3/4)
Remove obsolete audio_gain# blocks as their function is now integrated into rx_agc.
Connect uiDockAudio gain control to the rx_agc manual gain input.
Update uiDockAudio gain control with rx_acg current gain values in auto
modes.
UI and connection logic fixes to make everything work.

+++ src/dsp/agc_impl.cpp
@@ -71,7 +71,7 @@ CAgc::~CAgc()
 }
 
 void CAgc::SetParameters(double sample_rate, bool agc_on, int target_level,
-                         int manual_gain, int max_gain, int attack, int decay,
+                         float manual_gain, int max_gain, int attack, int decay,
                          int hang, bool force)
 {
     bool samp_rate_changed = false;
